Rename base busybox-debian to glibc-busybox-bash
The new image name reflects more what the image consist of (i.e., glibc,
busybox, and bash) than the implementation detail of how (i.e., using
Debian) it is built.
If the build image's base changes (e.g., moving to a RHEL-based distro)
in the future, that change should be reflected in a major version bump.